🚧 Flesh out the create post example a bit

+++ packages/api/src/router/post.ts
@@ -1,3 +1,4 @@
+import { TRPCError } from "@trpc/server";
 import { z } from "zod";
 
 import { createTRPCRouter, protectedProcedure, publicProcedure } from "../trpc";
@@ -11,14 +12,33 @@ export const postRouter = createTRPCRouter({
 		.query(({ ctx, input }) => {
 			return ctx.prisma.post.findFirst({ where: { id: input.id } });
 		}),
-	delete: publicProcedure.input(z.string()).mutation(({ ctx, input }) => {
-		return ctx.prisma.post.delete({ where: { id: input } });
-	}),
+	delete: protectedProcedure
+		.input(z.string())
+		.mutation(async ({ ctx, input }) => {
+			const post = await ctx.prisma.post.findFirst({
+				where: { id: input },
+				select: { userId: true },
+			});
+
+			if (!post)
+				throw new TRPCError({ code: "NOT_FOUND", message: "Post not found" });
+
+			const isAuthorizedToDelete =
+				ctx.auth.userId === post.userId || ctx.auth.orgRole === "admin";
+
+			if (!isAuthorizedToDelete)
+				throw new TRPCError({
+					code: "UNAUTHORIZED",
+					message: "You are not authorized to delete this post",
+				});
+
+			return ctx.prisma.post.delete({ where: { id: input } });
+		}),
 	create: protectedProcedure
 		.input(
 			z.object({
-				title: z.string(),
-				content: z.string(),
+				title: z.string().min(1, "Title must be at least 1 character"),
+				content: z.string().min(1, "Content must be at least 1 character"),
 			}),
 		)
 		.mutation(({ ctx, input }) => {
